Add files via upload
his is project is exploration as well as model development. Whereby, I explored the data on factors that can lead to diabetes. At the end, I also trained the model and the prediction accuracy is around 64%.

This accuracy is not satisfactory, thus, the accuracy will be pushed up in further changes in this pdf!